requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/LLM
openai-swarm==0.1.0
anthropic==0.7.8

# Database
aiosqlite==0.19.0

# Vercel Serverless
mangum==0.17.0

# Utilities
python-multipart==0.0.6
